Modified scraping methodology
Scraping methodology has been changed to allow for a wider array of websites to be scraped. Rather than manually making a specific scraper for the 500+ websites that the EDAMAM API utilizes, a more generic scraping approach has been implemented. We now look for directions/instructions/methods to know when to start scraping. When one of those are found we look for the next element to determine the expected element type of the instructions. While the next element types match the original child type we keep scraping. We stop scraping if the element type changes and has a string length greater than 30 (arbitrary number as of right now) or if we are at the reviews or comments section. Also, during scraping if the element matches the original element, but is less than 30 chars in length we don't add to the directions as it is more than likely a picture caption or image tag.